added geom_match to subpixel_register_point
